\name{metabias}
\alias{metabias}
\concept{test}
\title{Test for Funnel Plot Asymmetry}
\description{Test for funnel plot asymmetry.}
\usage{metabias(x, ...)}
\arguments{
  \item{x}{an object of class \code{"meta.nsue"}, or of other classes in other meta-analytic packages (e.g. \code{"meta"}).}
  \item{...}{other arguments.}
}
\details{The function \code{metabias} is generic. See \code{\link{metabias.meta.nsue}} for testing for funnel plot asymmetry for \code{"meta.nsue"} objects, or other methods from other meta-analytic packages (e.g. \code{\link[meta]{metabias.meta}}).}
\seealso{
  \code{\link{metabias.meta.nsue}} for testing for funnel plot asymmetry for \code{"meta.nsue"} objects.

  \code{\link[=funnel.meta.nsue]{funnel}} for plotting funnel plots.

  \code{\link[=meta.nsue]{meta}} for conducting a meta-analysis.

  \code{\link[=metalm.meta.nsue]{metalm}} for fitting meta-analytic models.
}
\keyword{bias}
\keyword{meta-analysis}
\keyword{regression}
\keyword{test}
